PLAT-278: Fix circular deps
